requirements.txt
django-uuslug==1.2.0
Pillow==8.1.0
six==1.15.0
